moved around the code to clean it up
