+++ src/test/java/pages/ListAddressPage.java
@@ -1,9 +1,9 @@
-package tests.pages;
+package pages;
 
 import com.saucelabs.framework.elements.Element;
 import com.saucelabs.framework.pages.OnPage;
 import org.openqa.selenium.By;
-import tests.data.AddressData;
+import data.AddressData;
 
 @OnPage(path="/addresses")
 public class ListAddressPage extends BasePage {
