+++ src/test/java/pages/LogInPage.java
@@ -1,10 +1,10 @@
-package tests.pages;
+package pages;
 
 import com.saucelabs.framework.elements.Element;
 import com.saucelabs.framework.elements.TextField;
 import com.saucelabs.framework.pages.OnPage;
 import org.openqa.selenium.By;
-import tests.data.UserData;
+import data.UserData;
 
 @OnPage(path="/sign_in")
 public class LogInPage extends BasePage {
